class Pair<K, V> -> inner class; edit field size
